requirements.txt
kafka-python==2.0.1
# kafka==2.0.1